support cascading of NLS from app down to nested views (#105). Some refactoring to levarege the same code in View & app. Some refactoring to move all utilities to utils.
