+++ tests/globalizedApp/home.html
@@ -1,5 +1,7 @@
 <div class="view mblView">
     <ul data-dojo-type="dojox/mobile/RoundRectList">
+		<li data-dojo-type="dojox/mobile/ListItem" data-dojo-props="'clickable':true, 'label': '${nls.label0}'">
+  		</li>
         <li data-dojo-type="dojox/mobile/ListItem" data-dojo-props="'clickable':true, 'label': '${nls.label1}'">
         </li>
         <li data-dojo-type="dojox/mobile/ListItem" data-dojo-props="'clickable':true">
