+++ tests/globalizedApp/nls/app.js
@@ -1,0 +1,6 @@
+define({
+  root: {
+   	label0: "Label Zero"
+  },
+  fr: true
+});
