+++ tests/globalizedApp/nls/fr/app.js
@@ -1,0 +1,6 @@
+define({
+  root: {
+   	label0: "Etiquette 0"
+  },
+  fr: true
+});
